add new component and add plain auto time keeper func
